move backend-independent stuff
