+++ plugins/KeyboardioHID/src/MultiReport/Mouse.h
@@ -29,7 +29,7 @@ THE SOFTWARE.
 #include <Arduino.h>
 #include "HID.h"
 #include "HID-Settings.h"
-#include "../MouseButtons.h"
+#include "kaleidoscope/MouseButtons.h"
 #include "tusb_hid.h"
 
 #define DESCRIPTOR_MOUSE(...)                          \
